[WIP] Replace axi_ArbitrationTree by axi_node_arbiter
- versions of `common_cells` (in commit message and `Bender.yml`) and
  `axi` (in `Bender.yml`)
- clarify FPGA area measurements

Instead of reimplementing round robin arbitration through a recursively
instantiated tree, `axi_node_arbiter` simply instantiates
`stream_arbiter`, which is a variant of the solid `rrarbiter` that is
compatible with stream protocols such as AXI.  Those modules are part of
`common_cells` (>= v1.TODO).

This commit fixes two crucial problems of `axi_node` as AXI
interconnect:

1. The round robin arbitration is now actually fair.  Previously, some
   masters were preferred in the arbitration (not to starvation, but the
   former arbitration had a bias of ca. 50% towards certain masters).

   A significant speed-up can be observed when equal distribution of
   work and data movement is important.  For example, mounting 4
   clusters with 8 cores simultaneously takes ca. 40% less cycles to
   execute.

2. Once the arbiter has decided on a master to forward, it keeps its
   decision.  Previously, the arbiter changed its decision after
   asserting the valid signal at the output, meaning that AXI data
   changed while valid was high, which violates the AXI standard.

   The former version of the interconnect required one AXI slice per
   master and slave port to prevent this violation of the standard.  If
   timing and clock domain constraints permit, these slices can now be
   removed.

Two use cases of `axi_node` as SoC bus, one on an ASIC and one on an
FPGA, have been investigated:  When used in UMCL65 with 5 master and 3
slave ports and without slices on any port, both the combinatorial and
non-combinatorial area of `soc_bus_wrap` increase by ca. 20%.  When used
on a Xilinx 7-Series FPGA with 10 master and 5 slave ports and with
slices on all ports, the number of LUTs and FFs used by `soc_bus_wrap`
decreases by 20% and 50%, respectively.

This version of `axi_node` is interface-compatible with its predecessor
with the exception that the dependency ordering between ready and valid
is now stricter.  Remember that the AXI standard [AXI4, Issue E, Sec.
A3.3.1] mandates that, in any transaction:

-  the VALID signal of the interface sending information MUST NOT depend
   on the READY signal of the interface receiving that information, and

-  an interface that is receiving information CAN wait until it detects
   a VALID signal before it asserts its corresponding READY signal.

Modules that violate at least one of these conditions may cause
deadlocks and should be fixed immediately.

This commit reduces the code complexity of `axi_node` and reuses simple
modules that can be verified in isolation.

`axi_node` now requires `common_cells` >= v1.TODO (previously
unspecified).